refactor: moved addUserVariables method to helpers
